Silence a "sparse" warning.
2006-03-16  Morten Welinder  <terra@gnome.org>

	* src/cell.c (cell_copy): Silence a "sparse" warning.

2006-03-16  Morten Welinder  <terra@gnome.org>

	* ms-excel-read.c (excel_read_EXTERNNAME): Use NULL, not 0.
	* ms-biff.c (ms_biff_query_next): Ditto.
	(ms_biff_query_destroy): Ditto.
	(ms_biff_put_new): Ditto.
	(ms_biff_put_len_next): Ditto.
	(ms_biff_put_len_commit): Ditto.

+++ src/cell.c
@@ -127,7 +127,9 @@ cell_copy (GnmCell const *cell)
 
 	/* The new cell is not linked into any of the major management structures */
 	new_cell->base.sheet = NULL;
-	new_cell->base.flags &= ~(DEPENDENT_NEEDS_RECALC|CELL_IN_SHEET_LIST|DEPENDENT_IS_LINKED);
+	new_cell->base.flags &= ~(DEPENDENT_NEEDS_RECALC |
+				  (int)CELL_IN_SHEET_LIST |
+				  DEPENDENT_IS_LINKED);
 
 	/* now copy properly the rest */
 	if (cell_has_expr (new_cell))
